Шапка и подвал формируются из шаблонов. Каталог и корзина заполняются динамически.
