+++ src/main/java/ru/zakharov/enterprise/dao/CartDAO.java
@@ -1,0 +1,4 @@
+package ru.zakharov.enterprise.dao;
+
+public class CartDAO {
+}
