+++ src/main/java/ru/zakharov/enterprise/dao/ProductDAO.java
@@ -1,0 +1,4 @@
+package ru.zakharov.enterprise.dao;
+
+public class ProductDAO {
+}
