+++ src/main/java/ru/zakharov/enterprise/entity/Product.java
@@ -1,0 +1,4 @@
+package ru.zakharov.enterprise.entity;
+
+public class Product {
+}
