+++ src/main/java/ru/zakharov/enterprise/servlet/CartServlet.java
@@ -1,16 +1,14 @@
 package ru.zakharov.enterprise;
 
 import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {"/product"})
-public class ProductServlet extends HttpServlet {
+public class CartServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("product.jsp").forward(req,resp);
+        req.getRequestDispatcher("WEB-INF/cart.jsp").forward(req,resp);
     }
 }
