+++ src/main/java/ru/zakharov/enterprise/servlet/CatalogServlet.java
@@ -6,9 +6,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class MainServlet extends HttpServlet {
+public class CatalogServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("index.jsp").forward(req,resp);
+        req.getRequestDispatcher("WEB-INF/catalog.jsp").forward(req,resp);
     }
 }
