+++ src/main/java/ru/zakharov/enterprise/servlet/Error403Handler.java
@@ -7,10 +7,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {"/error404"})
-public class Error404Handler extends HttpServlet {
+@WebServlet(urlPatterns = {"/error403"})
+public class Error403Handler extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("error404.jsp").forward(req, resp);
+        req.getRequestDispatcher("WEB-INF/error403.jsp").forward(req, resp);
     }
 }
