+++ src/main/java/ru/zakharov/enterprise/servlet/Error404Handler.java
@@ -7,10 +7,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {"/error403"})
-public class Error403Handler extends HttpServlet {
+@WebServlet(urlPatterns = {"/error404"})
+public class Error404Handler extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("error403.jsp").forward(req, resp);
+        req.getRequestDispatcher("WEB-INF/error404.jsp").forward(req, resp);
     }
 }
