+++ src/main/java/ru/zakharov/enterprise/servlet/MainServlet.java
@@ -6,9 +6,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class CartServlet extends HttpServlet {
+public class MainServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("cart.jsp").forward(req,resp);
+        req.getRequestDispatcher("WEB-INF/index.jsp").forward(req,resp);
     }
 }
