+++ src/main/webapp/WEB-INF/templates/footer.jsp
@@ -1,0 +1,16 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Alex
+  Date: 06.08.2018
+  Time: 22:38
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>$Title$</title>
+</head>
+<body>
+$END$
+</body>
+</html>
